+++ docs/changelog.rst
@@ -6,6 +6,65 @@ Release History
 
 .. towncrier release notes start
 
+Nasdaq_100_Ticker_History 2025.8.0 (2025-10-10)
+===============================================
+
+Internal Changes
+----------------
+- Move to uv-build and src layout for the project
+- Update python dependencies
+
+
+Features
+--------
+
+- Thompson Reuters TRI replaces Ansys ANSS after Ansys was acquired by Synopsys.
+
+
+Nasdaq_100_Ticker_History 2025.7.0 (2025-07-31)
+===============================================
+
+Features
+--------
+
+- Thompson Reuters TRI replaces Ansys ANSS after Ansys was acquired by Synopsys.
+
+
+Nasdaq_100_Ticker_History 2025.6.0 (2025-07-06)
+===============================================
+
+Features
+--------
+
+- Add support for calendar year 2015.
+
+
+Nasdaq_100_Ticker_History 2025.5.0 (2025-07-05)
+===============================================
+
+Internal Changes
+----------------
+
+- Refactor tests into per-year files for better organization.
+- Add a GitHub workflow to validate pull requests.
+- Update python dependencies.
+
+Bug Fixes
+---------
+
+- Correctly handle the ticker symbol rename of Trip.com (formerly CTRP) in 2019.
+
+
+Nasdaq_100_Ticker_History 2025.4.0 (2025-05-19)
+===============================================
+
+Features
+--------
+
+Update Nasdaq 100 constituents:
+  - Shopify (SHOP) replaced MongoDB (MDB) on 19 May.
+
+
 Nasdaq_100_Ticker_History 2025.2.0 (2024-12-25)
 ===============================================